[fix] fix instruction OP_LD

+++ src/cpu/execution/inst_32/base_i.rs
@@ -277,7 +277,7 @@ pub fn exec(inst: &Instruction, cpu: &mut Cpu) -> Result<(), (Option<u64>, TrapC
             let load_addr = cpu.trans_addr(
                 TransFor::Load,
                 TransAlign::Size64,
-                (cpu.regs.read(inst.rs1) as i32 + inst.imm.unwrap()) as u64,
+                (cpu.regs.read(inst.rs1) as i64 + inst.imm.unwrap() as i64) as u64,
             )?;
             cpu.regs.write(inst.rd, cpu.bus.load64(load_addr)?);
         }
